% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqcluster.R
\name{seqcluster}
\alias{seqcluster}
\title{Cluster sequences}
\usage{
seqcluster(seqdistance, h = 1.5, cmethod = "complete")
}
\arguments{
\item{seqdistance}{A distance matrix or a distance array returned by TraMineR::seqdist() function.}

\item{h}{Numeric scalar or vector with heights where the tree should be cut. The default is 1.5.}

\item{cmethod}{This argument will be passed to \code{hclust()}. It takes the same values as \code{method} argument in \code{hclust()}. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). The default clustering method is "complete", complete-linkage clustering.}
}
\value{
\code{seqcluster} returns a data frame with three columns—the index of original data, group memberships, and group size.
}
\description{
This function clusters sequences using hierarchical cluster and returns group memberships of sequences.
}
\examples{
data(mvad)
mvad.alphab <- c("employment", "FE", "HE", "joblessness", "school", "training")
mvad.seq <- seqdef(mvad, 17:86, xtstep = 6, alphabet = mvad.alphab)
seqdistance <- seqdist(mvad.seq, method="HAM")
clusters <- seqcluster(seqdistance, h=1.5, cmethod='complete')
}
